Move SNIPhase function to kubeapiserverexposure.go
